+++ src/components/common/timeSearch/TimeSearchFilters.vue
@@ -32,12 +32,10 @@
 					class="explanation-for-data"
 				>
 					<div>
-						Datamængde viser en kurve over data vi har i arkivet lorem ipsum dolor sit amet, consectetur adipiscing
-						elit, sed do eiusmod tempor incididunt ut.
+						{{ $t('search.dataQuantity.firstPart') }}
 					</div>
 					<div>
-						Der er huller i arkivet, men vi arbejder på det lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
-						do eiusmod tempor incididunt ut.
+						{{ $t('search.dataQuantity.secondPart') }}
 					</div>
 					<button
 						ref="closeDataExplanation"
@@ -347,7 +345,7 @@ export default defineComponent({
 					data.value.push({ key: i, value: i });
 				}
 				if (dataButton.value) {
-					dataButton.value.style.display = 'none';
+					dataButton.value.classList.remove('display-data-button');
 				}
 				if (searchResultStore.firstBackendFetchExecuted && Object.keys(searchResultStore.initFacets).length !== 0) {
 					constructSVG();
@@ -420,7 +418,7 @@ export default defineComponent({
 				dataContainer.value.appendChild(svgElement);
 			}
 			if (dataButton.value) {
-				dataButton.value.style.display = 'block';
+				dataButton.value.classList.add('display-data-button');
 			}
 		};
 
@@ -919,6 +917,10 @@ fieldset {
 	top: 7px;
 }
 
+.display-data-button {
+	display: none;
+}
+
 .day-button,
 .month-button,
 .time-button {
@@ -1066,6 +1068,10 @@ fieldset {
 		display: block;
 	}
 	.data-size {
+		display: block;
+	}
+
+	.display-data-button {
 		display: block;
 	}
 	.slider-container {