+++ src/components/global/content-elements/EdgedContentArea.vue
@@ -9,7 +9,12 @@
 			ref="content"
 			class="full-width"
 		>
-			<div class="container"><slot name="content"></slot></div>
+			<div
+				ref="container"
+				class="container"
+			>
+				<slot name="content"></slot>
+			</div>
 		</div>
 		<div
 			ref="bottom"
@@ -55,11 +60,18 @@ export default defineComponent({
 				return false;
 			},
 		},
+		alignItems: {
+			type: String as PropType<string>,
+			default() {
+				return 'center';
+			},
+		},
 	},
 
 	setup(props) {
 		const top = ref<HTMLDivElement>();
 		const content = ref<HTMLDivElement>();
+		const container = ref<HTMLDivElement>();
 		const bottom = ref<HTMLDivElement>();
 
 		const getClasses = (prefix: string) => {
@@ -73,10 +85,11 @@ export default defineComponent({
 		onMounted(() => {
 			if (top.value) top.value.style.backgroundColor = props.backgroundColor;
 			if (content.value) content.value.style.backgroundColor = props.backgroundColor;
+			if (container.value) container.value.style.alignItems = props.alignItems;
 			if (bottom.value) bottom.value.style.backgroundColor = props.backgroundColor;
 		});
 
-		return { top, content, bottom, getClasses };
+		return { top, content, container, bottom, getClasses };
 	},
 });
 </script>
@@ -171,8 +184,15 @@ h3.reverse {
 	margin-left: auto;
 	box-sizing: border-box;
 	display: flex;
-	align-items: center;
 	flex-direction: column;
+}
+
+/* MEDIA QUERY 450 */
+@media (min-width: 450px) {
+	h3 {
+		font: normal normal normal 49px/50px Libre Baskerville;
+		font-size: 40px;
+	}
 }
 /* MEDIA QUERY 480 */
 @media (min-width: 480px) {
@@ -184,6 +204,10 @@ h3.reverse {
 @media (min-width: 640px) {
 	.container {
 		max-width: 990px;
+	}
+	h3 {
+		font: normal normal normal 49px/68px Libre Baskerville;
+		font-size: 48px;
 	}
 }
 /* MEDIA QUERY 990 */