+++ src/components/global/nav/wc-header-menu.ts
@@ -337,7 +337,7 @@ const MENU_COMPONMENT_STYLES = /*css*/ `
 
 	.overall-header {
 		position:relative;
-		z-index:2;
+		z-index:4;
 	}
 
 	.search-label {
@@ -868,6 +868,7 @@ const MENU_COMPONMENT_STYLES = /*css*/ `
 		}
 
 		.overall-header {
+			z-index: 5;
 			margin-bottom:0px;
 		}
 		.container {